refactor: move onboarding orchestration out of repository client
