+++ src/commands/pull.ts
@@ -4,8 +4,8 @@ import { CommandError, createCommand, type CommandConfig } from "../lib/command"
 import { diffArrays } from "../lib/diff";
 import { getDirtyPaths, getGitRoot } from "../lib/git";
 import { getCustomTypes, getSlices } from "../lib/prismic/clients/custom-types";
-import { completeOnboardingSteps } from "../lib/prismic/clients/repository";
 import { canonicalizeCustomType, canonicalizeSlice } from "../lib/prismic/models";
+import { completeOnboardingSteps } from "../lib/prismic/onboarding";
 import { isDescendant, relativePathname } from "../lib/url";
 import { findProjectRoot, getRepositoryName } from "../project";
 
@@ -145,11 +145,10 @@ export default createCommand(config, async ({ values }) => {
 
 	await adapter.generateTypes();
 
-	await completeOnboardingSteps({
+	await completeOnboardingSteps(["connectPrismic"], {
 		repo: await getRepositoryName(),
 		token,
 		host,
-		stepIds: ["connectPrismic"],
 	}).catch(() => {});
 
 	const totalTypes = customTypeOps.insert.length + customTypeOps.update.length;
