+++ src/commands/repo-create.ts
@@ -4,8 +4,8 @@ import { detectAgent } from "../lib/ai";
 import { CommandError, createCommand, type CommandConfig } from "../lib/command";
 import { upsertLocale } from "../lib/prismic/clients/locale";
 import { activateMCP } from "../lib/prismic/clients/mcp";
-import { completeOnboardingSteps } from "../lib/prismic/clients/repository";
 import { checkIsDomainAvailable, createRepository } from "../lib/prismic/clients/wroom";
+import { completeOnboardingSteps } from "../lib/prismic/onboarding";
 
 const MAX_DOMAIN_TRIES = 5;
 
@@ -54,11 +54,10 @@ export async function createRepo(config: {
 	// A new repository has no locale, so set the master locale to make it usable.
 	await upsertLocale({ id: lang, isMaster: true }, { repo: domain, token, host });
 
-	await completeOnboardingSteps({
+	await completeOnboardingSteps(["createPrismicProject"], {
 		repo: domain,
 		token,
 		host,
-		stepIds: ["createPrismicProject"],
 	}).catch(() => {});
 
 	await activateMCP({ repo: domain, token, host }).catch(() => {});
