+++ src/commands/sync.ts
@@ -8,7 +8,7 @@ import { getErrorMessage } from "../error";
 import { createCommand, type CommandConfig, CommandError } from "../lib/command";
 import { diffArrays } from "../lib/diff";
 import { getCustomTypes, getSlices } from "../lib/prismic/clients/custom-types";
-import { completeOnboardingSteps } from "../lib/prismic/clients/repository";
+import { completeOnboardingSteps } from "../lib/prismic/onboarding";
 import { getRepositoryName } from "../project";
 import { trackCommandStart, trackCommandEnd } from "../tracking";
 
@@ -121,11 +121,10 @@ export default createCommand(config, async ({ values }) => {
 				lastHash = nextHash;
 
 				if (isInitial) {
-					await completeOnboardingSteps({
+					await completeOnboardingSteps(["connectPrismic"], {
 						repo: await getRepositoryName(),
 						token,
 						host,
-						stepIds: ["connectPrismic"],
 					}).catch(() => {});
 					console.info("Initial sync complete.");
 				} else {
