+++ src/lib/prismic/clients/repository.ts
@@ -23,44 +23,30 @@ export function getRepository(config: RepositoryConfig): Promise<Repository> {
 	return repositoryServiceRequest(url, config, { schema: RepositorySchema });
 }
 
-export type OnboardingStep =
-	| "createPrismicProject"
-	| "createPageType"
-	| "createSlice"
-	| "connectPrismic";
-
 const OnboardingStateSchema = z.object({
 	completedSteps: z.array(z.string()),
 });
 export type OnboardingState = z.infer<typeof OnboardingStateSchema>;
 
-type OnboardingConfig = RepositoryConfig;
-
-export async function getOnboardingState(config: OnboardingConfig): Promise<OnboardingState> {
+export async function getOnboardingState(config: RepositoryConfig): Promise<OnboardingState> {
 	const url = new URL("onboarding", getRepositoryServiceUrl(config.host));
 	return onboardingServiceRequest(url, config, {
 		schema: OnboardingStateSchema,
 	});
 }
 
-export async function completeOnboardingSteps(
-	config: OnboardingConfig & { stepIds: OnboardingStep[] },
-): Promise<void> {
-	const { host, stepIds } = config;
-	const { completedSteps } = await getOnboardingState(config);
-	const missing = stepIds.filter((id) => !completedSteps.includes(id));
-
-	// API does not accept multiple steps; toggle each missing step sequentially.
-	for (const stepId of missing) {
-		const url = new URL(
-			`onboarding/${encodeURIComponent(stepId)}/toggle`,
-			getRepositoryServiceUrl(host),
-		);
-		await onboardingServiceRequest(url, config, {
-			method: "PATCH",
-			schema: OnboardingStateSchema,
-		});
-	}
+export async function toggleOnboardingStep(
+	stepId: string,
+	config: RepositoryConfig,
+): Promise<OnboardingState> {
+	const url = new URL(
+		`onboarding/${encodeURIComponent(stepId)}/toggle`,
+		getRepositoryServiceUrl(config.host),
+	);
+	return onboardingServiceRequest(url, config, {
+		method: "PATCH",
+		schema: OnboardingStateSchema,
+	});
 }
 
 function repositoryServiceRequest<T>(
